refactor: typed SubmissionStatus/UserRole, NotebookWorkingCopyStore extraction
SubmissionStatus and UserRole raw-value enums replace the stringly-typed
literal comparisons across the server (DB columns unchanged; Fluent
filters use rawValue, plain comparisons use typed accessors).

The ~420-line notebook working-copy helper block moves out of
WebRoutes+Notebook.swift into Services/NotebookWorkingCopyStore.swift,
and the pre-v0.4 legacy notebook-copy sweep now runs once at boot
instead of on every notebook page view.
