+++ Sources/APIServer/Models/APISubmission.swift
@@ -3,6 +3,21 @@
 import Fluent
 import Vapor
 
+/// Lifecycle states for a submission row.
+///
+/// The `status` DB column stays a plain string (no migration); this enum is
+/// the authoritative vocabulary for it.  `running` is included because the
+/// diagnostics queries and the MCP validation watcher both recognize it as a
+/// live state, even though the server itself currently writes only the other
+/// four.
+enum SubmissionStatus: String, Sendable {
+    case pending
+    case assigned
+    case running
+    case complete
+    case failed
+}
+
 final class APISubmission: Model, Content, @unchecked Sendable {
     // @unchecked Sendable: all mutations happen within Vapor's request context,
     // never across unstructured concurrency.
@@ -20,7 +35,7 @@ final class APISubmission: Model, Content, @unchecked Sendable {
     var testSetupID: String
 
     @Field(key: "status")
-    var status: String  // pending | assigned | complete | failed
+    var status: String  // SubmissionStatus raw value; column stays a string
 
     @OptionalField(key: "worker_id")
     var workerID: String?
@@ -77,7 +92,7 @@ final class APISubmission: Model, Content, @unchecked Sendable {
         testSetupID: String,
         zipPath: String,
         attemptNumber: Int,
-        status: String = "pending",
+        status: String = SubmissionStatus.pending.rawValue,
         filename: String? = nil,
         userID: UUID? = nil,
         kind: String = Kind.student
@@ -90,5 +105,19 @@ final class APISubmission: Model, Content, @unchecked Sendable {
         self.filename = filename
         self.userID = userID
         self.kind = kind
+    }
+}
+
+// MARK: - Typed status accessors
+
+extension APISubmission {
+    /// The submission's status as a typed enum, or nil if the stored string
+    /// is outside the known vocabulary (defensive — should not happen).
+    var statusValue: SubmissionStatus? { SubmissionStatus(rawValue: status) }
+
+    /// Sets the status from the typed enum.  Prefer this over assigning a
+    /// raw string to `status`.
+    func setStatus(_ newStatus: SubmissionStatus) {
+        status = newStatus.rawValue
     }
 }
