+++ Sources/APIServer/Routes/Web/AdminRoutes+Courses.swift
@@ -437,7 +437,7 @@ extension AdminRoutes {
                 .filter(\.$id ~~ enrolledUserIDs)
                 // Exclude `mcp` service accounts: enrolled to scope an agent's
                 // access (admin MCP tab), not human roster members.
-                .filter(\.$role != "mcp")
+                .filter(\.$role != UserRole.mcp.rawValue)
                 .sort(\.$username)
                 .all()
             enrolledUsers = users.compactMap { u in
